apps/openai-cursor-transcriber/requirements.txt
openai>=1.0.0,<2
pyside6>=6.7.0,<7
python-dotenv>=1.0.0,<2
pynput>=1.7.6,<2
sounddevice>=0.4.6,<1
